% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclbl_functions.R
\name{save_bart}
\alias{save_bart}
\title{Generate Scailable BART model}
\usage{
save_bart(
  bart_model,
  filename = "bart.zip",
  upload = FALSE,
  sclbl_config_file = "sclbl_config.yml"
)
}
\arguments{
\item{bart_model}{fitted bart model}

\item{filename}{name of the generated zip file (default: "bart.zip")}

\item{upload}{whether to automatically upload the model to scailable.net or not (default: FALSE)}

\item{sclbl_config_file}{location of config file containing the scailable API endpoint URL and
scailable credentials (default: sclbl_config.yml)}
}
\description{
Takes a fitted BART model and generates a zip file
from it that can then be uploaded to scailable.net
}
\examples{

\dontrun{
  load("bart_model.RData")
  save_bart(bart_model, "bart.zip", upload = FALSE)
}

}
